Removed unused imports and some minor refactoring
