+++ src/main/java/net/sf/jabref/gui/actions/NewSubDatabaseAction.java
@@ -5,7 +5,6 @@ import net.sf.jabref.Globals;
 import net.sf.jabref.BibDatabaseContext;
 import net.sf.jabref.JabRefPreferences;
 import net.sf.jabref.gui.*;
-import net.sf.jabref.gui.util.PositionWindow;
 import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.model.database.BibDatabaseMode;
 import net.sf.jabref.gui.auximport.FromAuxDialog;
