+++ src/main/java/net/sf/jabref/gui/journals/ManageJournalsAction.java
@@ -17,7 +17,6 @@ package net.sf.jabref.gui.journals;
 
 import net.sf.jabref.gui.JabRefFrame;
 import net.sf.jabref.gui.actions.MnemonicAwareAction;
-import net.sf.jabref.gui.util.PositionWindow;
 import net.sf.jabref.logic.l10n.Localization;
 
 import java.awt.event.ActionEvent;
