+++ src/main/java/net/sf/jabref/importer/ImportCustomizationDialog.java
@@ -42,8 +42,6 @@ import net.sf.jabref.*;
 import net.sf.jabref.gui.FileDialogs;
 import net.sf.jabref.gui.GUIGlobals;
 import net.sf.jabref.gui.util.FocusRequester;
-import net.sf.jabref.gui.util.PositionWindow;
-
 import com.jgoodies.forms.builder.ButtonBarBuilder;
 
 /**
