+++ src/main/java/net/sf/jabref/sql/DbConnectAction.java
@@ -16,7 +16,6 @@
 package net.sf.jabref.sql;
 
 import net.sf.jabref.gui.actions.BaseAction;
-import net.sf.jabref.gui.util.PositionWindow;
 import net.sf.jabref.gui.BasePanel;
 import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.sql.exporter.DBExporter;
