+++ src/main/java/net/sf/jabref/sql/importer/DbImportAction.java
@@ -31,7 +31,6 @@ import org.apache.commons.logging.LogFactory;
 
 import net.sf.jabref.gui.*;
 import net.sf.jabref.gui.actions.MnemonicAwareAction;
-import net.sf.jabref.gui.util.PositionWindow;
 import net.sf.jabref.gui.worker.AbstractWorker;
 import net.sf.jabref.model.database.BibDatabase;
 import net.sf.jabref.Globals;
